Updated OpenWeatherService and UI

+++ network/src/main/java/com/example/network/model/geo/GeoCode.kt
@@ -7,11 +7,11 @@ data class GeoCode(
     @SerializedName("country")
     val country: String = "",
     @SerializedName("lat")
-    val lat: Float = 0.0f,
+    val lat: Double = 0.0,
     @SerializedName("local_names")
     val localNames: LocalNames = LocalNames(),
     @SerializedName("lon")
-    val lon: Float = 0.0f,
+    val lon: Double = 0.0,
     @SerializedName("name")
     val name: String = "",
     @SerializedName("state")
